requirements.txt
numpy
numpy-stl
open3d
opencv-contrib-python
opencv-python
pandas
Pillow
pymeshlab
pyntcloud
scikit-image
scipy
timm==0.4.5
torch==1.10.0
torchvision==0.11.0
tqdm
vtk
wandb
pycocotools
matplotlib
colorlog
shapely
pyqt5==5.14.2
